+++ CHANGES.rst
@@ -4,6 +4,7 @@ CHANGE HISTORY
 LATEST
 ------
 - Added border option to PopupMenu
+- Added scrolling of truncated text to ListBox and MulticolumnListBox.
 - Fixed double-width alignment issues on DropdownList
 - Improved TextBox line wrapping to break on word boundaries.
 - Fixed logic for highlighting selected widget controls without focus.